Move non-matcher code to _deferred
